requirements.txt
openai>=1.55.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
supabase>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

